requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.38.0
ftfy==6.1.1
idna==3.4
imageio==2.23.0
ipython==8.7.0
jedi==0.18.2
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
networkx==2.8.8
numpy==1.24.1
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
regex==2022.10.31
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
stack-data==0.6.2
tifffile==2022.10.10
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5

# package location
--find-links https://download.pytorch.org/whl/torch_stable.html